requirements.txt
torch==2.5.1+cpu
torchvision==0.20.1+cpu
torchaudio==2.5.1+cpu
numpy>=1.26.0
matplotlib>=3.8.0
pandas>=2.2.0
Pillow>=10.4.0
opencv-python>=4.9.0
scipy>=1.12.0
tqdm>=4.66.0
tensorboard>=2.16.0
PyYAML>=6.0.0